% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbsegdata-package.R
\docType{package}
\name{fafbsegdata-package}
\alias{fafbsegdata-package}
\alias{_PACKAGE}
\alias{fafbsegdata}
\title{fafbsegdata: Additional version data to accompany fafbseg data}
\description{
A package to distribute versioned support data for the fafbseg
package. Using a separate package for large data objects should avoid
repository churn.
}
\examples{
\donttest{
library(fafbseg)
find_merged_segments(7186840767)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/fafbsegdata}
  \item Report bugs at \url{https://github.com/jefferis/fafbsegdata/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{This}
\keyword{\code{\link{segment_merge_groups.dt}}}
\keyword{\code{fafbseg}}
\keyword{a}
\keyword{by}
\keyword{data}
\keyword{internal}
\keyword{object}
\keyword{only}
\keyword{package.}
\keyword{presently}
\keyword{provides}
\keyword{the}
\keyword{used}
