% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segment_merge_groups.dt}
\alias{segment_merge_groups.dt}
\title{Data object defining the merge groups for raw segments}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 17308654 rows and 2 columns.}
\usage{
segment_merge_groups.dt
}
\description{
Data object defining the merge groups for raw segments
}
\details{
The original information was shared by Peter Li in 10 csv files
}
\examples{
head(segment_merge_groups.dt)
}
\keyword{datasets}
